Fix `Copy` && `Clone` for ErasedFnPointer while removed Debug impl
As pointed out by
[here](https://stackoverflow.com/questions/60907129/rust-copy-trait-not-working-with-a-type-when-it-uses-generic),
using `#[derive(Copy, Clone)]` will gives an implementation of `Copy`
and `Clone` only works when the generic parameters all impl `Copy` and
`Clone`, so I have to implements these manually.

As for `Debug`, it also does the same thing, so I would rather remove it
than take my time and implements it.

Signed-off-by: Jiahao XU <Jiahao_XU@outlook.com>